pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
 
[project]
name = "uart_proxy"
version = "0.0.1"
authors = [
    { name="Felix Schladt", email="felix.schladt@hensoldt.net" },
]
description="uart proxy for TRENTOS OSS hardware ci"
requires-python=">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Private :: Do Not Upload",
]
dependencies=[
    "fastapi",
    "requests",
    "uvicorn",
    "pyserial",
    "python-multipart",
]
 
[project.scripts]
uart_proxy = "uart_proxy.main:main"
 
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
uart_proxy = ["*.json"]
